site/fargate/chat_ui/requirements.txt
langchain==0.0.189
pydantic==1.10.7
PyYaml
streamlit==1.22.0
streamlit-ace==0.1.1
streamlit-authenticator==0.2.2
streamlit-chat==0.0.2.2
typing-inspect==0.8.0
typing_extensions==4.3.0